Adding Some of the CSV  files I used
This is just a very small amount of the data that I used to set up the NN. I would like to say, after reasearch the action of using price for LTSM neural network development is wrong. There should be the use of Percents, and maybe some kind of Technical indicator to give the daily price fluctuations shape.